[#3] feat: viewModel 구현
