+++ C2_DiverDexApp/Views/Profile/CategoryTag.swift
@@ -1,0 +1,18 @@
+//
+//  CategoryTag.swift
+//  C2_DiverDexApp
+//
+//  Created by PenguinLand on 4/16/25.
+//
+
+import SwiftUI
+
+struct CategoryTag: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    CategoryTag()
+}
