+++ C2_DiverDexApp/Views/Profile/InfoRowView.swift
@@ -1,0 +1,8 @@
+//
+//  InfoRowView.swift
+//  C2_DiverDexApp
+//
+//  Created by PenguinLand on 4/16/25.
+//
+
+import Foundation
